requirements.txt
snowflake-connector-python
snowflake-snowpark-python
python-dotenv
pyyaml
pydantic
pydantic-settings
